1. Fixed Spellings in README.md 2. Added comment in BugDatabaseProcessor.cs 3. Removed extra checks in BugDatabaseProcessor.ProcessBugDatabase 4. Moved interfaces into their own files 5. BugDatabaseDllLoader.Load now throws instead of returning null 6. Updated tests to accommodate changes
