+++ src/vcsparser.bugdatabase.azuredevops/IApiConverter.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using vcsparser.core.bugdatabase;
+
+namespace vcsparser.bugdatabase.azuredevops
+{
+    public interface IApiConverter
+    {
+        WorkItem ConvertToWorkItem(dynamic fullWorkItem);
+    }
+}
