+++ src/vcsparser.core/bugdatabase/IHttpClientWrapper.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net.Http;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace vcsparser.core.bugdatabase
+{
+    public interface IHttpClientWrapper
+    {
+        Task<HttpResponseMessage> SendAsync(HttpRequestMessage request);
+    }
+}
