+++ src/vcsparser.core/bugdatabase/TimeKeeper.cs
@@ -7,15 +7,6 @@ using System.Threading.Tasks;
 
 namespace vcsparser.core.bugdatabase
 {
-    public interface ITimeKeeper
-    {
-        TimeSpan Delay { get; set; }
-        Action IntervalAction { get; set; }
-        bool IsCompleted { get; }
-        Task Start();
-        void Cancel();
-    }
-
     public class TimeKeeper : ITimeKeeper
     {
         public TimeSpan Delay { get; set; }
