+++ src/vcsparser.core/vcsparser.core.csproj
@@ -58,12 +58,19 @@
     <Compile Include="bugdatabase\BugDatabaseFactory.cs" />
     <Compile Include="bugdatabase\HttpClientWrapper.cs" />
     <Compile Include="bugdatabase\HttpClientWrapperFactory.cs" />
+    <Compile Include="bugdatabase\IBugDatabaseDllLoader.cs" />
+    <Compile Include="bugdatabase\IBugDatabaseFactory.cs" />
+    <Compile Include="bugdatabase\IBugDatabaseProcessor.cs" />
     <Compile Include="bugdatabase\IBugDatabaseProvider.cs" />
+    <Compile Include="bugdatabase\IHttpClientWrapper.cs" />
+    <Compile Include="bugdatabase\IHttpClientWrapperFactory.cs" />
+    <Compile Include="bugdatabase\ITimeKeeper.cs" />
+    <Compile Include="bugdatabase\IWebRequest.cs" />
     <Compile Include="bugdatabase\TimeKeeper.cs" />
     <Compile Include="bugdatabase\WebRequest.cs" />
     <Compile Include="bugdatabase\WorkItem.cs" />
     <Compile Include="bugdatabase\BugDatabaseProcessor.cs" />
-    <Compile Include="bugdatabase\WorkItemConverter.cs" />
+    <Compile Include="bugdatabase\IWorkItemConverter.cs" />
     <Compile Include="ChangesetProcessor.cs" />
     <Compile Include="DailyCodeChurnAuthor.cs" />
     <Compile Include="DailyCodeChurnBugDatabase.cs" />
